pyproject.toml
[tool.poetry]
name = "LipoBART"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9,<3.11"
catboost = "^1.2.2"
numpy = "^1.26.3"
matplotlib = "^3.8.2"
pandas = "^2.1.4"
scikit-learn = "^1.3.2"
seaborn = "^0.13.1"
deepchem = "^2.7.1"
xgboost = "^2.0.3"
ipykernel = "^6.29.0"
hydra-core = "^1.3.2"

[tool.poetry.group.dev.dependencies]
pylint = "^3.0.2"
black = "^23.11.0"
isort = "^5.12.0"
pre-commit = "^3.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
